Move UserException to its own file
